fix: Refactor to remove circular import
This commit refactors the code to eliminate a circular import between `__init__.py` and `sensor.py`.

The `LenedaApiClient` class has been moved from `__init__.py` to a new `api.py` file. Both `__init__.py` and `sensor.py` now import the API client from `api.py`.

This should resolve the silent "Failed to set up" error.

Bumps the integration version to 0.1.5.
