+++ custom_components/leneda/manifest.json
@@ -7,5 +7,5 @@
   "codeowners": [],
   "requirements": [],
   "iot_class": "cloud_polling",
-  "version": "0.1.4"
+  "version": "0.1.5"
 }
